"Chapter Three"

+++ chapterTwo/questionSeven.scala
@@ -1,4 +1,4 @@
 //Solve the preceding exercise without writing a loop. (Hint: Look at the StringOps
 //Scaladoc.)
 
-println("Hello".foldRight(1)((a,b) => a.toInt * b.toInt))
+println("Hello".foldRight(1)((a,b) => a * b))
